requirements.txt
alembic==1.7.6
click==8.0.3
Flask==2.0.2
Flask_Bootstrap==3.3.7.1
Flask_Login==0.5.0
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.0
Pillow==9.0.1
requests==2.27.1
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
